bigos: remove the chunk notion from __memory_move
Remove the possibility to have moved chunks with a size > 4Kb.
This is because we only want to move some pages basing on their usage
so the notion of chunk is irrelevant.
Moreover, there is no really "out" thing and removing different chunk
size possibility between "in" and "out" remove a LOT of useless
complexity.