bigos: add argument to __memory_move() to indicate where to allocate
Add a parameter to the __memory_move() function to indicate, for each
specified gfn, on what node to allocate the replacement page.
If the node is unexisting, it is considered as full and the
replacement page cannot be allocated.
In this case, the execution flow continue as for normal failure
allocation.