bigos: disable PEBS support
Forbid any use of PEBS because of the following problem:
PEBS use linear address to indicate where it stores the samlings.
When entering in guest mode, PEBS use the guest page table to
translate this linear address in physical address.
This translation can make the guest crash, or even corrupts ithe guest
data.