bigos: fix a bug on the memory move
During the memory move, the moved source page was write-protected to
avoid modification during the copy to its destination page.
To write-protect the source page, the p2m_access_rx was applied to it.
However, in Xen, this is a soft state which can be reset to Xen
default access (p2m_access_rwx).

Now the type of the page is changed to p2m_ram_ro which is a hard
state. This type is indicated to silently drop writes, but we need to
catch these writes to make the accessing guest waiting the page is
moved.
Luckily, this silent drop is done in the page fault handler of the
VMEXIT, so we just have to catch the fault before it is dropped.