bigos: fix a lock ordering bug
The increasing of sampling rate revealed a bug which used to happen
rarely: under some circumstances, the interrupt code can try to
acquire the p2m lock whereas it already has a higher order lock which
is forbidden (to avoid deadlocks).

This commit make the p2m_read_trylock() function check if a higher
order lock is already taken, and if so, fail to lock.