bigos: fix a race condition for IBS
Fix a race condition used to happen while sampling a core performing
an asynchronous context switch. The guest page table was read during
a write_cr3 leading to a kernel crash.

Now when an asynchronous context switch is detected, the IBS handler
exit instead of reading the guest page table.