bigos: add a storage for IBS samples in the Carrefour rbtree
Add a call in the Carrefour IBS handler to store the samples in the
Carrefour rbtree.
One of the arguments of this call is a boolean indicating kernel mode
(in the original Carrefour).
It is now replaced by a boolean indicating hypervisor mode, but to
have this information, it is necessary to bring down the cpu_user_regs
from the do_nmi() function to the final handler.