bigos: fix bugs on krealloc() bridge and on sampling
Fix a bug on the carrefour alloc krealloc() function. The size of the
old area was incorrect and the memcpy() call was making triple fault.

Fix a bug on IBS sampling making it think it was always in hypervisor
context. Not yet found how to do it properly, so say it is never
hypervisor context.