bigos: fix a race condition on IBS
Before the fix, an NMI interrupt could be triggered during the IBS
disabling and especially during the xenoprof_arch_stop(). This could
lead to:
- interrupt triggered
- xenoprof_arch_stop set the IBS counters to 0
- the nmi_ibs() handler does no see anything in IBS counter
- the nmi_ibs(à handler return 0 whereas this is an IBS interrupt
- an unknown interrupt is detected and a BUG() is called

Now the ibs_disable() uses an asynchronous disabling mechansim
ensuring the hardware IBS is disabled at the end of an NMI interrupt.
The ibs_disable() still returns only when no more IBS interrup can be
triggered.