bigos: remove the irq enabling during IBS guest page walk
Before this commit, the guest walk in IBS handling would require irq
enabling for avoiding the Xen locks crying.
This was a short-term fix leading to deadlocks anyway since irq should
not be enabled in an NMI context.

Now, the entire call tree of the guest walk is modified to take an new
abort parameter, indicating if it is true, the functions should use
read_trylock instead of lock.

The trylock is a lock not testing if irq is enabled or not but which
cannot block (since it is a trylock) and thus, avoiding the issue
described in check_lock().