bigos: fix a performance problem in ibs_disable
IBS can only be disabled in an event-driven way, because of
xenoprofile: once ibs_disable() is called, it waits until every code
make an IBS interrupt.
When sampling happens at a low rate, didabling IBS can be very costly.
This patch increase the sampling rate before to start the disabling
process to reduce the overhead.